Upgraded sensor values to doubles

+++ 7701V/include/robot-config.h
@@ -37,13 +37,13 @@ extern rotation sRot;
 
 extern inertial Inertial;
 
-extern int redMin, redMax;
-extern int blueMin, blueMax;
+extern double redMin, redMax;
+extern double blueMin, blueMax;
 extern optical leftRollerOptical;
 extern optical rightRollerOptical;
 
 extern int matchLoadStartDelay, matchLoadEndDelay;
-extern int storageDistMin, storageDistMax;
+extern double storageDistMin, storageDistMax;
 extern distance storageDistance;
 
 extern const uint32_t shotCooldown;
